corrected teh datatype of score field to have decimal value
Former-commit-id: b8559b6e92bcae4eafec59e2b26c7bdd6c2625f4
Former-commit-id: 92234c8ea3407c1b6986b7f1a5f400474264de59

+++ scripts/database/mosip_regprc/ddl/regprc-abis_response_det.sql
@@ -13,7 +13,7 @@
 CREATE TABLE regprc.abis_response_det(
 	abis_resp_id character varying(36) NOT NULL,
 	matched_bio_ref_id character varying(36) NOT NULL,
-	score smallint NOT NULL,
+	score numeric(6,3) NOT NULL,
 	cr_by character varying(256) NOT NULL,
 	cr_dtimes timestamp NOT NULL,
 	upd_by character varying(256),
